Cleaned up a lot of the code, added message throttling.
Added scope guards to ensure that we don't leak memory and allows us to
reduce code duplication.
Moved utility functions out of the main.cpp file.
Added a throttler to ensure that we don't send more than 20 messages
every 30 seconds (since that will result in twitch banning the bot for 2
hours automatically). Capped at 20 since we aren't 100% guaranteed to be
OP/ mod status on the channel we are joining (which would make the cap
100 instead).