Paying Tech Debt and adding user commands.
Throller updated to be constructed with a timeout instead of using a
    hard-coded 30s.
utils updated for typo
Makefile updates to use ag instead of find.
README updated to reflect ag requirement for building.
main updated to add member list and points (1 point per minute of being
    in the chat).
Added json submodule to allow us to save point values off to a file
    easily. Should at some point be updated to use sqlite instead.