Got CAP REQs working! Updated member list and added some new functions.
Working towards a raffle system in the chat. This will probably be put
on hold until we can come up with a better way to handle asynchronous
messaging within the application since I am not liking the jury-rigged
approach we ware currently taking (see GETROLE/RETURNROLE for more
datails on this).