Added additional CAP REQs and fixed non-blocking accept
Turns out that we weren't setting up the non-blocking accept checks
appropriately and as a result we were seeing blocking accepts that we
weren't expecting.

Added 2 new cap requests to the bot so we now get commands and tags as
well as membership. This should allow up to determine when someone
follows/ subs a stream as well as seeing banned/ user state information.

Cleaned up some of the code.

Fixed the shutdown!