Consolidated some lambdas for various threads.
Transceiver now is completely alone in having access to the twitch IRC
connection. It is created, connected, and cleaned up all withing this
thread.

API server is now also consolidated with the API loop thread. Due to no
longer being blocking it should be able to easily deal with accepting
and handling requests without needing the separate thread.